Move WeightInfo, Event into Settlement pallet.
