+++ pallets/common/src/traits/mod.rs
@@ -26,4 +26,3 @@ pub mod asset;
 pub mod balances;
 pub mod checkpoint;
 pub mod compliance_manager;
-pub mod settlement;
